refactor(core/rdma): keep core/ strictly device-safe
Move the host-only *_dv.h shims (bnxt_re_dv/ionic_dv/mlx5_dv — they pull system
verbs/mlx5dv and are used only by application/) out of core/ into
application/transport/rdma/providers/. Drop the dead host helpers
to_ionic_mr_flags/ionic_to_ibv_status from ionic_fw.h (no callers) so the
firmware ABI header no longer carries <infiniband/verbs.h>; add <stdint.h> for
the uint*_t it uses directly.
