+++ include/mori/application/transport/rdma/providers/ionic/ionic.hpp
@@ -26,9 +26,9 @@
 #include <hsa/hsa_ext_amd.h>
 
 #include "mori/application/transport/rdma/providers/dv_loader.hpp"
+#include "mori/application/transport/rdma/providers/ionic/ionic_dv.h"
 #include "mori/application/transport/rdma/rdma.hpp"
 #include "mori/core/transport/rdma/providers/ionic/ionic_defs.hpp"
-#include "mori/core/transport/rdma/providers/ionic/ionic_dv.h"
 #include "mori/core/transport/rdma/providers/ionic/ionic_fw.h"
 
 namespace mori {
